refactor spec workflow
